changed name to begin with test, added imports, added unittesting
